Remove faulty error type checks.
These type checks were not doing anything - which was surfaced by a `go vet` run on the more recent
Go version. errors.As with the second argument of type *error will always return true.

I played a little bit with trying to get a working typecheck in this test, but it didn't really work
because of issues with the fork github.com/Clever/syslogparser. Since that parser didn't update
self-references to point to itself instead of the upstream, our errors returned by this are already
not really working correctly - sometimes they are Clever/syslogparser.ParseError and sometimes
jeromer/syslogparser.ParseError. Fixing the underlying problem is out of scope for right now.